+++ __mocks__/node-ssh.js
@@ -16,13 +16,20 @@ module.exports = class NodeSsh {
    * execCommand mock
    */
   execCommand (command) {
+    let result
+
     if (command.includes('connection-error')) {
-      return {
-        stderr: 'Error occured'
-      }
+      result = { stderr: 'Error occured' }
+    } else if (command.includes('.limit(1)')) {
+      if (command.includes('.skip(0)')) {
+        result = { stdout: '[{"id": 1}]' }
+      } else if (command.includes('.skip(1)')) {
+        result = { stdout: '[{"id": 2}]' }
+      } else result = { stdout: '[]' }
+    } else {
+      result = { stdout: '[{"id": 1}, {"id": 2}]' }
     }
-    return {
-      stdout: '[]'
-    }
+
+    return result
   }
 }