+++ CHANGELOG.md
@@ -20,7 +20,11 @@
 - [ ] Improve the jsdoc comments so that a jsdoc wesite report looks good
 - [ ] Improve readme with complete documentation
 - [ ] CodeImprovement: all functions take only a object as input
+- [ ] Clean up tests
 - [x] ~~implement recurring sync~~ (moved to the mojoin server, since it's only needed there)
+
+- Error handling for:
+  - [ ] idField does not exist
 
 ## Version 0.0.7 - 2018-xx-xx
 
@@ -30,8 +34,8 @@
 
 ## Version 0.0.6 - 2018-xx-xx
 
-- [ ] Datasource query pagination
-- [ ] Logging library debug
+- [x] Datasource query pagination
+- [X] Logging library debug
 - [x] Insert projects logo in the readme
 
 ## Version 0.0.5 - 2018-10-11